Added D07P01, continued working on D06P02

+++ AdventOfCode.Shared/AocDayPartResult.h
@@ -2,5 +2,5 @@
 
 struct AocDayPartResult
 {
-    int result;
+    int64_t result;
 };
